+++ NEWS.md
@@ -1,4 +1,10 @@
+# tectonicr 0.3.7 _2024_06_09_
+
+* download WSM2016 data from GFZ server using `load_WSM2016()`
+
 # tectonicr 0.3.6
+
+* bug fixes in (weighted) rayleigh test
 
 # tectonicr 0.3.5